# docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CreateGroupRequest:
    properties:
      name:
        type: string
    type: object
  dto.CreateGroupResponse:
    properties:
      id:
        type: integer
    type: object
  dto.CreateStudentRequest:
    properties:
      group_id:
        type: integer
      name:
        type: string
    type: object
  dto.CreateStudentResponse:
    properties:
      id:
        type: integer
    type: object
...
  dto.DeleteGroupResponse:
    properties:
      message:
        type: string
    type: object
  dto.DeleteStudentResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.GetGroupsResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/dto.GroupDto'
        type: array
    type: object
  dto.GetStudentsGroupResponse:
    properties:
      students:
        items:
          $ref: '#/definitions/dto.StudentDto'
        type: array
    type: object
...
  dto.GroupDto:
    properties:
      id:
        type: integer
      name:
        type: string
      students:
        items:
          $ref: '#/definitions/dto.StudentDto'
        type: array
    type: object
...
  dto.StudentDto:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  dto.UpdateGroupRequest:
    properties:
      name:
        type: string
    type: object
  dto.UpdateGroupResponse:
    properties:
      message:
        type: string
    type: object
  dto.UpdateStudentRequest:
    properties:
      group_id:
        type: integer
      name:
        type: string
    type: object
  dto.UpdateStudentResponse:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для генерации фидбека учителям.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Feedback Teachers Helper API
  version: "1.0"
paths:
...
  /groups:
    get:
      description: Returns all groups for current teacher
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetGroupsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: List groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a group for current teacher
      parameters:
      - description: Group payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Create group
      tags:
      - groups
  /groups/{id}:
    delete:
      description: Deletes group by group id
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete group
      tags:
      - groups
    put:
      consumes:
      - application/json
      description: Updates group name by group id
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Update group
      tags:
      - groups
...
  /students:
    post:
      consumes:
      - application/json
      description: Creates student in a group
      parameters:
      - description: Student payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStudentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateStudentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Create student
      tags:
      - students
  /students/{groupId}:
    get:
      description: Returns students for group by group id
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetStudentsGroupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: List group students
      tags:
      - students
  /students/{id}:
    delete:
      description: Deletes student by student id
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteStudentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete student
      tags:
      - students
    put:
      consumes:
      - application/json
      description: Updates student by student id
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      - description: Student payload
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateStudentResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Update student
      tags:
      - students
...
securityDefinitions:
  Bearer:
    description: 'Введите токен в формате: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"